Load history from JHist files, not the history server
Timberlake currently eagerly loads data for past jobs from the
history server, the load from which can cause it to be sad sometimes.
This bypasses the historyserver to load job details, prefering to
load them directly from the .jhist files in HDFS.

This also contains a bunch of opportunistic cleanup and restructuring.